feat: adjust header actions ui response
